% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BloodPressure.R
\docType{data}
\name{BloodPressure}
\alias{BloodPressure}
\title{BloodPressure}
\format{dataframe}
\usage{
data(BloodPressure)
}
\arguments{
\item{Group}{Specifying the group of an individual. ("ctrl" = control, "case" = case)}

\item{Blood_pressure}{The measured bloodpressure of an individual.}
}
\description{
The bloodpressure is measured for each individual, either a control ("ctrl") or case ("case").
}
\details{
The BloodPressure dataset is a testset for the permutation(). It has one column specifying the groups (case or control) and another column with the observations (bloodpressure) for each individual in each group.
}
\keyword{datasets}
