% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_a_name.R
\name{draw_a_name}
\alias{draw_a_name}
\title{Plot the evolution of the numbers of time per year a name was given since 1900}
\usage{
draw_a_name(the_name, the_sex)
}
\arguments{
\item{the_name}{the name to plot}

\item{the_sex}{the sex related to the name}
}
\value{
a plot
}
\description{
Plot the evolution of the numbers of time per year a name was given since 1900
}
\examples{

draw_a_name(the_name = "Vincent",the_sex = "M")

}
