% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_names.R
\name{draw_names}
\alias{draw_names}
\title{Plot a graph comparing the evolutions of the frequency of two given names}
\usage{
draw_names(names)
}
\arguments{
\item{names}{the two names to compare}
}
\value{
a plot
}
\description{
Plot a graph comparing the evolutions of the frequency of two given names
}
\examples{
draw_names(c("Diane","Vincent"))
}
