% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Save a data set as a csv file}
\usage{
save_as_csv(dataset, filename, row.names = FALSE, ...)
}
\arguments{
\item{dataset}{dataset to save as csv}

\item{filename}{filename to save under}

\item{...}{other potential variables}
}
\value{
csv file
}
\description{
Save a data set as a csv file
}
